Process doubles instead of integers.

+++ app/src/main/java/com/interpreter/operators/Add.kt
@@ -10,7 +10,7 @@ class Add : Expression {
         rightExpression = right
     }
 
-    override fun interpret(): Int {
+    override fun interpret(): Double {
         return leftExpression.interpret() + rightExpression.interpret()
     }
